Major site redesign and additions
This commit adds some additional features, and improves upon existing design. Firstly, the two seam carving talks were added to the research page. Second, book reviews were added to reading.html. Book reviews can now be toggled to expand or collapse upon clicking a number next to the list entry. The number is the rating of the book out of 5. Pale green background for good books, pale red for the bad ones, and normal pale yellow for the rest. Also, toggling is controlled now by a universal function written in Javascript at the beginning of both pages in reading and research.